requirements.txt
Flask
gradio
openai
python-dotenv
